Apply `black` on all code - No manual adjustments
Highlighting the fact that I did not adjust anything.
There are some situations in which I thought about it and where the code might
have been written differently, if the writer would have known about the
adjustments that black makes (or just about the fact that lines are now 88
characters long), but overall just *blindly* applying it seems to be the most
consistent thing to do.

We might be interested to perform manual changes to further improve the code
style, but this can also be done once the PR is opened / in review.