Introduced `curr_rv`, `pred_rv` and `filt_rv`; `curr_rv` stays fixed
Before, `current` was modified during the step. I chose to have `curr_rv` stay
fixed as the random variable, from which we want to extrapolate to the next time
step, and instead loop over "predictions" (now `pred_rv`).

One nice thing is that we don't need to "reconstruct" this RV from the list of
means and covars anymore if a step was rejected.